pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "python-jose[cryptography]>=3.4.0",
    "streamlit>=1.42.2",
]
